src/requirements.txt
python-telegram-bot
localstack
boto3
boto3-type-annotations
pytest
pyyaml
